+++ StripeCryptoOnramp/StripeCryptoOnrampTests/Helpers/CryptoOnrampMockData.swift
@@ -65,6 +65,24 @@ enum ConfirmUserAttestationResponseMock: String, MockData {
     case confirmUserAttestationResponse_200 = "ConfirmUserAttestationResponse_200"
 }
 
+enum RetrievePartnerTermsResponseMock: String, MockData {
+    var bundle: Bundle { return Bundle(for: ClassForBundle.self) }
+
+    typealias ResponseType = PartnerTerms
+
+    case retrievePartnerTermsResponse_TermsAndConditionsRequired_200 = "RetrievePartnerTermsResponse_TermsAndConditionsRequired_200"
+    case retrievePartnerTermsResponse_TOSRequired_200 = "RetrievePartnerTermsResponse_TOSRequired_200"
+    case retrievePartnerTermsResponse_NotRequired_200 = "RetrievePartnerTermsResponse_NotRequired_200"
+}
+
+enum ConfirmPartnerTermsResponseMock: String, MockData {
+    var bundle: Bundle { return Bundle(for: ClassForBundle.self) }
+
+    typealias ResponseType = EmptyResponse
+
+    case confirmPartnerTermsResponse_200 = "ConfirmPartnerTermsResponse_200"
+}
+
 enum WalletOwnershipChallengeResponseMock: String, MockData {
     var bundle: Bundle { return Bundle(for: ClassForBundle.self) }